requirements.txt
Flask==3.1.2
flask_monitoringdashboard==4.0.4
keras==3.11.3
numpy==2.3.5
Pillow==12.0.0
pytest==8.4.2
Requests==2.32.5
scikit_learn==1.8.0
Werkzeug==3.1.4
